@@ -1,5 +1,4 @@
-const fs = require('fs')
+import { copyFileSync, mkdirSync } from 'fs';
 
-fs.copyFileSync('.env.sample', '.env')
-
-fs.mkdirSync('tmp/pgdata', { recursive: true })
+copyFileSync('.env.sample', '.env');
+mkdirSync('tmp/pgdata', { recursive: true });
