@@ -1,11 +1,16 @@
-import dotenv from 'dotenv'
-dotenv.config()
+import dotenv from 'dotenv';
+
+dotenv.config();
 
 export default {
   preset: 'ts-jest',
   testEnvironment: 'jsdom',
   testMatch: ['<rootDir>/src/**/*.test.{ts,tsx}'],
+  moduleNameMapper: {
+    '^@/(.*)$': '<rootDir>/src/$1',
+    '^.+\\.(css|scss|png|svg|jpg)$': 'babel-jest',
+  },
   globals: {
     __SERVER_PORT__: process.env.SERVER_PORT,
   },
-}
+};
