@@ -1,1 +1,1 @@
-declare const __SERVER_PORT__: number
+declare const __SERVER_PORT__: number;
