@@ -1,14 +1,16 @@
-const magic = '🪄'
+import { test, expect } from '@jest/globals';
 
-const cast = (spell: string, item: any) => {
+const magic = '🪄';
+
+const cast = (spell: string, item: string) => {
   if (spell.startsWith(magic)) {
-    return '🐷'
+    return '🐷';
   }
 
-  return item
-}
+  return item;
+};
 
 test('spell casting', () => {
-  const result = cast(magic, '🐸')
-  expect(result).toBe('🐷')
-})
+  const result = cast(magic, '🐸');
+  expect(result).toBe('🐷');
+});
