@@ -1,7 +1,8 @@
-import { Client } from 'pg'
+import { Client } from 'pg';
 
-const { POSTGRES_USER, POSTGRES_PASSWORD, POSTGRES_DB, POSTGRES_PORT } =
-  process.env
+const {
+  POSTGRES_USER, POSTGRES_PASSWORD, POSTGRES_DB, POSTGRES_PORT,
+} = process.env;
 
 export const createClientAndConnect = async (): Promise<Client | null> => {
   try {
@@ -11,18 +12,22 @@ export const createClientAndConnect = async (): Promise<Client | null> => {
       database: POSTGRES_DB,
       password: POSTGRES_PASSWORD,
       port: Number(POSTGRES_PORT),
-    })
+    });
 
-    await client.connect()
+    await client.connect();
 
-    const res = await client.query('SELECT NOW()')
-    console.log('  ➜ 🎸 Connected to the database at:', res?.rows?.[0].now)
-    client.end()
+    const res = await client.query('SELECT NOW()');
+    console.log('  ➜ 🎸 Connected to the database at:', res?.rows?.[0].now);
+    client.end();
 
-    return client
+    return client;
   } catch (e) {
-    console.error(e)
+    console.error(e);
   }
 
-  return null
-}
+  return null;
+};
+
+export default {
+  createClientAndConnect,
+};
